feat: integrate axum for improved web handling and update delegation TTL in documentation
