+++ Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 anyhow = "1.0.100"
+axum = "0.7.9"
 candid = "0.10.20"
 clap = { version = "4.5.51", features = ["derive"] }
 hex = "0.4.3"
